opt: Parenthesis optimization
Removing parentheses at the ends of lines is permissible in TI-BASIC, so we attempt to maximize the number of parentheses that will be at the end of every line.

Some operators have `opposites` which would produce the same result if the operator's LHS was swapped with its RHS (eg. `<=` and `>` are opposites, and `+` is its own opposite). We recursively count and swap so that each binary operator has more exposed closing parenthesis on its right hand side than its left.